[リファクタリング] MainControllerからWebSocketSharpの参照を取り除く
 MainControllerの中でWebSocketSharpを使っていた部分をWebSocketInitializerに隔離
 MainControllerがWebSocketInitializerをAwakeで見つけないとエラーが出る
