+++ Client/Assets/RollingBall.unity
@@ -734,6 +734,7 @@ GameObject:
   - component: {fileID: 2134770747}
   - component: {fileID: 2134770746}
   - component: {fileID: 2134770748}
+  - component: {fileID: 2134770749}
   m_Layer: 0
   m_Name: Main
   m_TagString: Untagged
@@ -752,7 +753,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 435317b1b875f40a688fdbd80cd25c61, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  connectAddress: ws://10.0.75.1:5678
   playerPrefab: {fileID: 1000010242128840, guid: b7ee01bb7113a428d83cd4c2b2dcff8f,
     type: 2}
   otherPlayerPrefab: {fileID: 1000013530691402, guid: 8da97395e89154d00aeaf2c9abf021e2,
@@ -782,3 +782,15 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: ecf68e73fa37649aa83b0295847cc7f6, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &2134770749
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 2134770745}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a7cc3511b2079084d9dcf69b7b0f027b, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  connectAddress: 10.0.75.1:5678
