+++ Client/Assets/Scripts/MainController.cs
@@ -1,16 +1,10 @@
 ﻿using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.SceneManagement;
-using WebSocketSharp;
 using RPC = WebSocketSample.RPC;
 
 public class MainController : MonoBehaviour
 {
-    WebSocket webSocket;    // WebSocketコネクション
-
-    [SerializeField]
-    string connectAddress;
-
     [SerializeField]
     GameObject playerPrefab;
     [SerializeField]
@@ -24,84 +18,11 @@ public class MainController : MonoBehaviour
     Dictionary<int, GameObject> otherPlayerObjs = new Dictionary<int, GameObject>();
     Dictionary<int, GameObject> items = new Dictionary<int, GameObject>();
 
-    void Start()
+    WebSocketInitializer socketInitializer;
+
+    void Awake()
     {
-        webSocket = new WebSocket(connectAddress);
-
-        // コネクションを確立したときのハンドラ
-        webSocket.OnOpen += (sender, eventArgs) =>
-        {
-            Debug.Log("WebSocket Opened");
-        };
-
-        // エラーが発生したときのハンドラ
-        webSocket.OnError += (sender, eventArgs) =>
-        {
-            Debug.Log("WebSocket Error Message: " + eventArgs.Message);
-        };
-
-        // コネクションを閉じたときのハンドラ
-        webSocket.OnClose += (sender, eventArgs) =>
-        {
-            Debug.Log("WebSocket Closed");
-        };
-
-        // メッセージを受信したときのハンドラ
-        webSocket.OnMessage += (sender, eventArgs) => 
-        {
-            Debug.Log("WebSocket Message: " + eventArgs.Data);
-
-            var header = JsonUtility.FromJson<RPC.Header>(eventArgs.Data);
-            switch (header.Method)
-            {
-                case "ping":
-                    {
-                        var pong = JsonUtility.FromJson<RPC.Ping>(eventArgs.Data);
-                        Debug.Log(pong.Payload.Message);
-                        break;
-                    }
-                case "login_response":
-                    {
-                        var loginResponse = JsonUtility.FromJson<RPC.LoginResponse>(eventArgs.Data);
-                        MainThreadExecutor.Enqueue(() => OnLoginResponse(loginResponse.Payload));
-                        break;
-                    }
-                case "sync":
-                    {
-                        var syncMessage = JsonUtility.FromJson<RPC.Sync>(eventArgs.Data);
-                        MainThreadExecutor.Enqueue(() => OnSync(syncMessage.Payload));
-                        break;
-                    }
-                case "spawn":
-                    {
-                        var spawnResponse = JsonUtility.FromJson<RPC.Spawn>(eventArgs.Data);
-                        MainThreadExecutor.Enqueue(() => OnSpawn(spawnResponse.Payload));
-                        break;
-                    }
-                case "delete_item":
-                    {
-                        var deleteMessage = JsonUtility.FromJson<RPC.DeleteItem>(eventArgs.Data);
-                        MainThreadExecutor.Enqueue(() => OnDeleteItem(deleteMessage.Payload));
-                        break;
-                    }
-                case "environment":
-                    {
-                        var environmentMessage = JsonUtility.FromJson<RPC.Environment>(eventArgs.Data);
-                        MainThreadExecutor.Enqueue(() => OnEnvironment(environmentMessage.Payload));
-                        break;
-                    }
-                case "delete_player":
-                    {
-                        var deletePlayerMessage = JsonUtility.FromJson<RPC.DeletePlayer>(eventArgs.Data);
-                        MainThreadExecutor.Enqueue(() => OnDeletePlayer(deletePlayerMessage.Payload));
-                        break;
-                    }
-            }
-        };
-
-        webSocket.Connect();
-
-        Login();
+        socketInitializer = GetComponent<WebSocketInitializer>();
     }
 
     void Update()
@@ -109,21 +30,16 @@ public class MainController : MonoBehaviour
         UpdatePosition();
     }
 
-    void OnDestroy()
-    {
-        webSocket.Close();    
-    }
-
-    void Login()
+    public void Login()
     {
         var jsonMessage = JsonUtility.ToJson(new RPC.Login(new RPC.LoginPayload("PlayerName")));
         Debug.Log(jsonMessage);
 
-        webSocket.Send(jsonMessage);
+        socketInitializer.Send(jsonMessage);
         Debug.Log(">> Login");
     }
 
-    void OnLoginResponse(RPC.LoginResponsePayload response)
+    public void OnLoginResponse(RPC.LoginResponsePayload response)
     {
         Debug.Log("<< LoginResponse");
         playerId = response.Id;
@@ -135,7 +51,7 @@ public class MainController : MonoBehaviour
         {
             var collisionRpc = new RPC.Collision(new RPC.CollisionPayload(playerId, otherPlayerId));
             var collisionJson = JsonUtility.ToJson(collisionRpc);
-            webSocket.Send(collisionJson);
+            socketInitializer.Send(collisionJson);
         };
     }
 
@@ -153,10 +69,10 @@ public class MainController : MonoBehaviour
         var rpcPosition = new RPC.Position(currentPlayerPosition.x, currentPlayerPosition.y, currentPlayerPosition.z);
         var jsonMessage = JsonUtility.ToJson(new RPC.PlayerUpdate(new RPC.PlayerUpdatePayload(playerId, rpcPosition)));
         Debug.Log(jsonMessage);
-        webSocket.Send(jsonMessage);
+        socketInitializer.Send(jsonMessage);
     }
 
-    void OnSync(RPC.SyncPayload payload)
+    public void OnSync(RPC.SyncPayload payload)
     {
         Debug.Log("<< Sync");
         foreach (var rpcPlayer in payload.Players)
@@ -187,7 +103,7 @@ public class MainController : MonoBehaviour
         }
     }
 
-    void OnSpawn(RPC.SpawnPayload payload)
+    public void OnSpawn(RPC.SpawnPayload payload)
     {
         Debug.Log("<< OnSpawn");
         SpawnItem(payload.Item);
@@ -207,12 +123,12 @@ public class MainController : MonoBehaviour
 
             var getItemRpc = new RPC.GetItem(new RPC.GetItemPayload(rpcItem.Id, playerId));
             var getItemJson = JsonUtility.ToJson(getItemRpc);
-            webSocket.Send(getItemJson);
+            socketInitializer.Send(getItemJson);
             Debug.Log(">> GetItem");
         };
     }
 
-    void OnDeleteItem(RPC.DeleteItemPayload payload)
+    public void OnDeleteItem(RPC.DeleteItemPayload payload)
     {
         Debug.Log("<< DeleteItem");
         var itemId = payload.ItemId;
@@ -223,7 +139,7 @@ public class MainController : MonoBehaviour
         }
     }
 
-    void OnEnvironment(RPC.EnvironmentPayload payload)
+    public void OnEnvironment(RPC.EnvironmentPayload payload)
     {
         Debug.Log("<< Environment");
 
@@ -250,7 +166,7 @@ public class MainController : MonoBehaviour
         }
     }
 
-    void OnDeletePlayer(RPC.DeletePlayerPayload payload)
+    public void OnDeletePlayer(RPC.DeletePlayerPayload payload)
     {
         if (otherPlayerObjs.ContainsKey(payload.Id))
         {
@@ -271,7 +187,7 @@ public class MainController : MonoBehaviour
 
     void RestartGame()
     {
-        webSocket.Close();
+        socketInitializer.Close();
         MainThreadExecutor.Clear();
         SceneManager.LoadScene(SceneManager.GetActiveScene().name);
     }
