+++ Client/Assets/Scripts/WebSocketInitializer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a7cc3511b2079084d9dcf69b7b0f027b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
